refactor(dashboard): move analytics route tracker for better organization
